move tests to snark_store_impl.rs
