+++ rust/tests/mod.rs
@@ -7,7 +7,6 @@ mod ledger;
 mod protocol;
 mod snark_work;
 mod state;
-mod store;
 mod usernames;
 
 pub mod helpers {
